refactor: move tables atoms in subfolder and fix table classnames
