+++ packages/core-app-elements/src/ui/atoms/tables/index.tsx
@@ -1,0 +1,5 @@
+import Tr from './Tr'
+import Th from './Th'
+import Td from './Td'
+
+export { Tr, Th, Td }
