+++ packages/core-app-elements/src/ui/tables/Table.tsx
@@ -6,10 +6,9 @@ interface Props extends React.TableHTMLAttributes<HTMLTableElement> {
   tbody?: React.ReactNode
 }
 
-function Table({ thead, tbody, ...rest }: Props): JSX.Element {
-  const topBorder = thead == null && 'border-t border-gray-100'
+function Table({ thead, tbody, className, ...rest }: Props): JSX.Element {
   return (
-    <table className={cn(['w-full', topBorder])} {...rest}>
+    <table className={cn(['w-full', className])} {...rest}>
       {thead != null && <thead>{thead}</thead>}
       {tbody != null && <tbody>{tbody}</tbody>}
     </table>
